pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lab2fhir"
version = "0.1.0"
description = "Unlock structured insights from your lab reports"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "Lab2FHIR Contributors"}
]
dependencies = [
    "pypdf>=3.17.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "fhir.resources>=7.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
lab2fhir = "lab2fhir.cli:main"

[tool.setuptools]
packages = ["lab2fhir"]
package-dir = {"" = "src"}

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"
